requirements.txt
# yaml~=0.2.5
pyyaml~=6.0
seaborn~=0.11.2
joblib>=1.2.0
scikit-learn~=1.0.1
numpy>=1.21.2
pandas
matplotlib
gradio
streamlit
google.generativeai
python-dotenv